refactor(Toast): change and separate filenames
