+++ src/components/Toaster/index.ts
@@ -1,0 +1,1 @@
+export { toast, ToastProps, Toaster } from './Toaster';
